feat: phase-aware landing hub + tabbed page reorganization
Replace the original Player Efficiency homepage with a landing hub that
surfaces quick-access cards grouped by area (Explore / Projections / Value /
Offseason), driven by the season-cycle resolver: a "Right now" phase banner +
deadline countdown, a featured-for-this-phase section, and sign-in cards for
gated groups when logged out. Since it's the pre-keeper phase, arbitration is
de-emphasized into a collapsed "Offseason" nav group and not featured.

Consolidate related pages into tabbed routes via a new server-rendered Tabs
component (active-only, URL-synced via ?tab=) + a client TabBar:
- /players gains Directory + Efficiency tabs (old / scatter folded in)
- /value merges VORP, Surplus, and Adjustments
- /arbitration merges Targets, Simulation, and Planner

Old URLs (/vorp, /surplus-value, /surplus-adjustments, /arbitration-simulation,
/arbitration-planner) redirect to the new tabs; middleware PROTECTED_ROUTES,
Navigation groups, and PHASE_UI featured links updated accordingly.

Surplus/Targets tables now build their function-bearing columns inside client
wrappers (SurplusTables, ArbTargetsTable) — the app's standard pattern — since
function props can't cross the server→client boundary.
