+++ web/app/arbitration/page.tsx
@@ -1,77 +1,20 @@
-import {
-  fetchPlayersWithProjectedPpg,
-  fetchAndMergeProjectedData,
-  fetchHoverExtras,
-  buildHoverDataMap,
-  analyzeArbitration,
-  allocateArbitrationBudget,
-  fetchPlayersPreArb,
-  ARB_BUDGET_PER_TEAM,
-  ARB_MIN_PER_TEAM,
-  ARB_MAX_PER_TEAM,
-  ARB_MAX_PER_PLAYER_PER_TEAM,
-  NUM_TEAMS,
-  LEAGUE_ID,
-  getHistoricalSeasonsForYear,
-} from "@/lib/analysis";
-import { getStatsSeason, getProjectionSeason } from "@/lib/season";
-import type { ArbitrationTarget, Column, HighlightRule } from "@/lib/types";
-import { getSupabaseAdmin } from "@/lib/supabase";
-import { getAuthenticatedUser } from "@/lib/auth";
-import DataTable from "@/components/DataTable";
-import ArbitrationTeams from "./ArbitrationTeams";
-import ActiveModelCard from "@/components/ActiveModelCard";
-import ModeToggle, { ValueMode } from "@/components/ModeToggle";
-import {
-  playerNameCol,
-  positionCol,
-  nflTeamCol,
-  ownerCol,
-  salaryCol,
-  valueCol,
-} from "@/components/columns";
+import Tabs from "@/components/Tabs";
+import type { ValueMode } from "@/components/ModeToggle";
+import TargetsSection from "./TargetsSection";
+import SimulationSection from "./SimulationSection";
+import PlannerSection from "./PlannerSection";
 
-interface Props {
-  searchParams: Promise<{ mode?: string }>;
-}
+// Per-user adjustments/plans are read here, so keep this route always-fresh.
+export const revalidate = 0;
 
-type ProjectedTarget = ArbitrationTarget & {
-  observed_ppg?: number;
+export const metadata = {
+  title: "Arbitration | Ottoneu Analytics",
+  description: "Arbitration targets, simulation, and planner for League 309",
 };
 
-function buildBaseColumns(hoverDataMap: Record<string, import("@/lib/types").PlayerHoverData> | null): Column<ProjectedTarget>[] {
-  return [
-    playerNameCol<ProjectedTarget>({ hoverDataMap }),
-    positionCol<ProjectedTarget>(),
-    nflTeamCol<ProjectedTarget>(),
-    ownerCol<ProjectedTarget>(),
-    salaryCol<ProjectedTarget>(),
-    valueCol<ProjectedTarget>(),
-    { key: "surplus", label: "Surplus", format: "currency" },
-    { key: "salary_after_arb", label: "After Arb", format: "currency" },
-    { key: "surplus_after_arb", label: "Surplus (Post-Arb)", format: "currency" },
-  ];
+interface Props {
+  searchParams: Promise<{ mode?: string; tab?: string }>;
 }
-
-function buildProjectedColumns(hoverDataMap: Record<string, import("@/lib/types").PlayerHoverData> | null): Column<ProjectedTarget>[] {
-  return [
-    playerNameCol<ProjectedTarget>({ hoverDataMap }),
-    positionCol<ProjectedTarget>(),
-    nflTeamCol<ProjectedTarget>(),
-    ownerCol<ProjectedTarget>(),
-    salaryCol<ProjectedTarget>(),
-    { key: "observed_ppg", label: "Obs PPG", format: "decimal" },
-    { key: "ppg", label: "Proj PPG", format: "decimal" },
-    valueCol<ProjectedTarget>(),
-    { key: "surplus", label: "Surplus", format: "currency" },
-    { key: "salary_after_arb", label: "After Arb", format: "currency" },
-    { key: "surplus_after_arb", label: "Surplus (Post-Arb)", format: "currency" },
-  ];
-}
-
-const ARB_TARGET_RULES: HighlightRule<ProjectedTarget>[] = [
-  { key: "surplus_after_arb", op: "lt", value: 0, className: "bg-red-50 dark:bg-red-950/30" },
-];
 
 export default async function ArbitrationPage({ searchParams }: Props) {
   const params = await searchParams;
@@ -81,202 +24,28 @@ export default async function ArbitrationPage({ searchParams }: Props) {
       : params.mode === "projected"
         ? "projected"
         : "raw";
-  const isProjected = mode === "projected";
-  const isAdjusted = mode === "adjusted";
-
-  // Fetch adjustments in all modes (needed for indicator dot)
-  const [user, statsSeason, projectionSeason] = await Promise.all([
-    getAuthenticatedUser(),
-    getStatsSeason(),
-    getProjectionSeason(),
-  ]);
-  const adjRes = user
-    ? await getSupabaseAdmin()
-        .from("surplus_adjustments")
-        .select("player_id, adjustment")
-        .eq("league_id", LEAGUE_ID)
-        .eq("user_id", user.userId)
-        .neq("adjustment", 0)
-    : { data: [], error: null };
-
-  const hasAdjustments = (adjRes.data?.length ?? 0) > 0;
-
-  let adjustments: Map<string, number> | undefined;
-  if (isAdjusted && adjRes.data && adjRes.data.length > 0) {
-    adjustments = new Map(
-      adjRes.data.map((r) => [String(r.player_id), Number(r.adjustment)])
-    );
-  }
-
-  // Fetch players with pre-arbitration salaries (after auto bump, before arb results)
-  let allPlayers;
-  if (isProjected) {
-    allPlayers = await fetchAndMergeProjectedData(projectionSeason, fetchPlayersPreArb);
-  } else {
-    // Raw mode uses projected PPG as the base
-    allPlayers = await fetchPlayersWithProjectedPpg(fetchPlayersPreArb);
-  }
-
-  const { projMap, dsMap } = await fetchHoverExtras(!!user?.hasProjectionsAccess);
-  const hoverDataMap = buildHoverDataMap(allPlayers, projMap, dsMap);
-
-  const targets = analyzeArbitration(allPlayers, adjustments) as ProjectedTarget[];
-
-  if (targets.length === 0) {
-    return (
-      <main className="min-h-screen bg-white dark:bg-black p-8">
-        <div className="max-w-7xl mx-auto">
-          <h1 className="text-3xl font-bold text-slate-900 dark:text-white">
-            No arbitration targets found.
-          </h1>
-        </div>
-      </main>
-    );
-  }
-
-  const sortedTeams = allocateArbitrationBudget(targets);
-
-  // When projected, augment team players with PPG data
-  const teamsData = isProjected
-    ? (() => {
-      const targetsLookup = new Map(
-        targets.map((t) => [`${t.name}|${t.team_name}`, t])
-      );
-      return sortedTeams.map(({ team, suggested, players }) => ({
-        team,
-        suggested,
-        players: players.map((p) => {
-          const full = targetsLookup.get(`${p.name}|${team}`);
-          return {
-            ...p,
-            observed_ppg: full?.observed_ppg ?? 0,
-            ppg: full?.ppg ?? 0,
-          };
-        }),
-      }));
-    })()
-    : sortedTeams;
-
-  const projectionYear = projectionSeason;
-  const historicalSeasons = getHistoricalSeasonsForYear(projectionYear);
-  const mostRecentSeason = Math.max(...historicalSeasons);
-  const columns = isProjected ? buildProjectedColumns(hoverDataMap) : buildBaseColumns(hoverDataMap);
 
   return (
     <main className="min-h-screen bg-white dark:bg-black p-8">
-      <div className="max-w-7xl mx-auto space-y-8">
+      <div className="max-w-7xl mx-auto space-y-6">
         <header>
-          <div className="flex flex-wrap items-start justify-between gap-4">
-            <div>
-              <h1 className="text-3xl font-bold tracking-tight text-slate-900 dark:text-white">
-                Arbitration Targets ({isProjected ? projectionYear : statsSeason})
-              </h1>
-              <p className="text-slate-500 dark:text-slate-400 mt-2">
-                {isProjected ? (
-                  <>
-                    Value based on <strong>projected future PPG</strong> rather than
-                    observed {mostRecentSeason} performance. Use this to target
-                    players whose value differs meaningfully from their recent stats.
-                  </>
-                ) : (
-                  <>
-                    Opponents&apos; players most vulnerable to a ${ARB_MAX_PER_PLAYER_PER_TEAM}{" "}
-                    arbitration raise. Negative surplus after arb = likely cut.
-                  </>
-                )}
-              </p>
-            </div>
-            <ModeToggle
-              currentMode={mode}
-              basePath="/arbitration"
-              hasAdjustments={hasAdjustments}
-            />
-          </div>
-          {isAdjusted && hasAdjustments && (
-            <div className="mt-3 bg-blue-50 dark:bg-blue-950/30 border border-blue-200 dark:border-blue-800 rounded-lg px-4 py-2 text-sm text-blue-800 dark:text-blue-300">
-              Showing results with your manual surplus adjustments applied.
-            </div>
-          )}
+          <h1 className="text-3xl font-bold tracking-tight text-slate-900 dark:text-white">
+            Arbitration
+          </h1>
+          <p className="text-slate-500 dark:text-slate-400 mt-2">
+            Targets, Monte Carlo simulation, and budget planning for the offseason
+            arbitration phase.
+          </p>
         </header>
 
-        {/* Projection Methodology — driven by projection_models.is_active */}
-        {isProjected && (
-          <ActiveModelCard
-            variant="blue"
-            footer={
-              <p className="text-xs text-blue-800/80 dark:text-blue-300/80 pt-1">
-                Built from {historicalSeasons.join(", ")} history. College
-                prospects fall back to positional rookie PPG averages.
-              </p>
-            }
-          />
-        )}
-
-        {/* Budget Info */}
-        <div className="bg-slate-50 dark:bg-slate-900 rounded-lg p-5 border border-slate-200 dark:border-slate-800">
-          <h2 className="text-lg font-semibold text-slate-900 dark:text-white mb-2">
-            Arbitration Budget
-          </h2>
-          <div className="grid grid-cols-2 sm:grid-cols-4 gap-4 text-sm">
-            <div>
-              <p className="text-slate-500 dark:text-slate-400">Total Budget</p>
-              <p className="font-bold text-slate-900 dark:text-white">
-                ${ARB_BUDGET_PER_TEAM}
-              </p>
-            </div>
-            <div>
-              <p className="text-slate-500 dark:text-slate-400">Per Team Range</p>
-              <p className="font-bold text-slate-900 dark:text-white">
-                ${ARB_MIN_PER_TEAM}-${ARB_MAX_PER_TEAM}
-              </p>
-            </div>
-            <div>
-              <p className="text-slate-500 dark:text-slate-400">
-                Max Per Player (from you)
-              </p>
-              <p className="font-bold text-slate-900 dark:text-white">
-                ${ARB_MAX_PER_PLAYER_PER_TEAM}
-              </p>
-            </div>
-            <div>
-              <p className="text-slate-500 dark:text-slate-400">Opponents</p>
-              <p className="font-bold text-slate-900 dark:text-white">
-                {NUM_TEAMS - 1}
-              </p>
-            </div>
-          </div>
-        </div>
-
-        {/* Top 20 Targets */}
-        <section>
-          <h2 className="text-xl font-semibold text-slate-900 dark:text-white mb-3">
-            Top 20 Arbitration Targets
-          </h2>
-          {isProjected && (
-            <p className="text-sm text-slate-500 dark:text-slate-400 mb-3">
-              <strong>Obs PPG</strong> = actual {mostRecentSeason}{" "}
-              season. <strong>Proj PPG</strong> = recency-weighted projection.
-              Red rows = negative projected surplus after a $
-              {ARB_MAX_PER_PLAYER_PER_TEAM} raise.
-            </p>
-          )}
-          <DataTable
-            columns={columns}
-            data={targets.slice(0, 20)}
-            highlightRules={ARB_TARGET_RULES}
-          />
-        </section>
-
-        {/* Per-Team Sections */}
-        <section>
-          <h2 className="text-xl font-semibold text-slate-900 dark:text-white mb-3">
-            Targets by Opponent
-          </h2>
-          <p className="text-sm text-slate-500 dark:text-slate-400 mb-4">
-            Suggested allocation based on {isProjected ? "projected" : "number of vulnerable"} surplus per team.
-          </p>
-          <ArbitrationTeams teams={teamsData} showProjectionColumns={isProjected} hoverDataMap={hoverDataMap} />
-        </section>
+        <Tabs
+          activeId={params.tab}
+          tabs={[
+            { id: "targets", label: "Targets", content: <TargetsSection mode={mode} /> },
+            { id: "simulation", label: "Simulation", content: <SimulationSection mode={mode} /> },
+            { id: "planner", label: "Planner", content: <PlannerSection /> },
+          ]}
+        />
       </div>
     </main>
   );
