+++ web/app/players/page.tsx
@@ -1,5 +1,8 @@
-import { fetchPlayerList } from "@/lib/data";
+import { fetchPlayerList, fetchPlayers } from "@/lib/data";
 import PlayerSearch from "@/components/PlayerSearch";
+import PlayerEfficiencyClient from "@/components/PlayerEfficiencyClient";
+import Tabs from "@/components/Tabs";
+import { ChartPoint } from "@/lib/types";
 
 export const revalidate = 3600; // Revalidate every hour
 
@@ -8,23 +11,105 @@ export const metadata = {
     description: "Browse and search all players in Ottoneu League 309",
 };
 
-export default async function PlayersPage() {
-    const players = await fetchPlayerList();
+/** Build the salary-vs-production scatter points (skips zero-PPG players). */
+function buildEfficiencyData(players: Awaited<ReturnType<typeof fetchPlayers>>): ChartPoint[] {
+    return players
+        .map((player) => {
+            const ppg = player.ppg;
+            const pps = player.pps;
+            const price = player.price;
+            const total_points = player.total_points;
+
+            if (ppg === 0) return null;
+
+            return {
+                name: player.name,
+                position: player.position,
+                nfl_team: player.nfl_team,
+                total_points,
+                ppg,
+                pps,
+                price,
+                cost_per_ppg: ppg > 0 ? price / ppg : 0,
+                cost_per_pps: pps > 0 ? price / pps : 0,
+                games_played: player.games_played,
+                snaps: player.snaps,
+            };
+        })
+        .filter(Boolean) as ChartPoint[];
+}
+
+interface Props {
+    searchParams: Promise<{ tab?: string }>;
+}
+
+export default async function PlayersPage({ searchParams }: Props) {
+    const { tab } = await searchParams;
+    const [players, efficiencyPlayers] = await Promise.all([
+        fetchPlayerList(),
+        fetchPlayers(),
+    ]);
+    const efficiencyData = buildEfficiencyData(efficiencyPlayers);
+
+    const directory = (
+        <div className="space-y-6">
+            <header>
+                <h2 className="text-2xl font-bold tracking-tight text-slate-900 dark:text-white">
+                    Player Directory
+                </h2>
+                <p className="text-slate-500 dark:text-slate-400 mt-2">
+                    Browse all players in League 309. Click a name to view their full
+                    card with stats and transaction history.
+                </p>
+            </header>
+            <PlayerSearch players={players} />
+        </div>
+    );
+
+    const efficiency = (
+        <div className="space-y-8">
+            <header>
+                <h2 className="text-2xl font-bold tracking-tight text-slate-900 dark:text-white">
+                    Player Efficiency
+                </h2>
+                <p className="text-slate-500 dark:text-slate-400 mt-2">
+                    Salary vs. production (Points Per Game or Points Per Snap).
+                </p>
+            </header>
+
+            <PlayerEfficiencyClient data={efficiencyData} />
+
+            <section className="bg-slate-50 dark:bg-slate-900 rounded-lg p-6 border border-slate-200 dark:border-slate-800">
+                <h3 className="text-lg font-semibold mb-4 text-slate-900 dark:text-white">Analysis Notes</h3>
+                <ul className="list-disc list-inside space-y-2 text-slate-600 dark:text-slate-400">
+                    <li><strong>Y-Axis (Salary)</strong>: Higher is more expensive.</li>
+                    <li><strong>X-Axis (PPG/PPS)</strong>: Further right means more efficient production per game/snap.</li>
+                    <li><strong>Bubble Size</strong>: Represents Total Points. Larger bubbles = higher total volume.</li>
+                    <li>Ideally, you want players in the <strong>Bottom-Right</strong> quadrant (High Production, Low Salary).</li>
+                </ul>
+            </section>
+        </div>
+    );
 
     return (
         <main className="min-h-screen bg-white dark:bg-black p-8">
             <div className="max-w-7xl mx-auto space-y-6">
                 <header>
                     <h1 className="text-3xl font-bold tracking-tight text-slate-900 dark:text-white">
-                        Player Directory
+                        Players
                     </h1>
                     <p className="text-slate-500 dark:text-slate-400 mt-2">
-                        Browse all players in League 309. Click a name to view their full
-                        card with stats and transaction history.
+                        Search the league directory or explore the salary-vs-production chart.
                     </p>
                 </header>
 
-                <PlayerSearch players={players} />
+                <Tabs
+                    activeId={tab}
+                    tabs={[
+                        { id: "directory", label: "Directory", content: directory },
+                        { id: "efficiency", label: "Efficiency", content: efficiency },
+                    ]}
+                />
             </div>
         </main>
     );
